+++ src/redux/reducers/users.js
@@ -1,0 +1,13 @@
+// 2019-10-19
+// Daniel Nicolas Gisolfi
+
+import * as ActionTypes from '../constants';
+import { initialState } from './index';
+
+export const userReducer = (state = initialState, action) => {
+	switch (action.type) {
+
+		default:
+			return state;
+	}
+};
